Rework Broadcaster methods to always need an encoding and do the encoding themselves
This way, all broadcasted messages are always garanteed to have an encoding.
Also, properly document methods and sanitize messages.

For now, '<' '>' characters aren't escaped yet, but that could easily be added.

Finally, suggested by @dbaeta, since the global broadcastToPlayers is basically
the same as the limited one, we can just run the latter with the complete set
of playerList nicks to indicate the limited one to send to all the players.